refactor(garm): move desired-entity building into a charm_state module
Introduce src/charm_state.py with a dataclasses-based CharmState
(CharmState.from_charm), per ISD014 and the pattern in #251, and move
the entity builder, the credential-name helper, and the configurator
relation-name constant there. charm.py builds entities via
CharmState.from_charm(self).desired_entities.

Also trim the _reconcile_runners docstring to an interface description
and move the step-ordering rationale to an inline comment, and note in
AGENTS.md that garm now uses a dataclass CharmState.
