+++ charms/garm/tests/unit/test_charm.py
@@ -848,7 +848,6 @@ def test_reconcile_runners_reconciles_credentials_entities_then_scalesets():
     entities = [object()]
     scalesets = [object()]
     charm._build_desired_credentials = MagicMock(return_value=credentials)
-    charm._build_desired_entities = MagicMock(return_value=entities)
     charm._build_desired_scalesets = MagicMock(return_value=scalesets)
     charm._ensure_controller_urls = MagicMock()
     charm.restart = MagicMock()
@@ -858,8 +857,10 @@ def test_reconcile_runners_reconciles_credentials_entities_then_scalesets():
         patch("charm.GithubReconciler") as mock_github_cls,
         patch("charm.EntityReconciler") as mock_entity_cls,
         patch("charm.ScalesetReconciler") as mock_scaleset_cls,
+        patch("charm.CharmState") as mock_charm_state_cls,
         patch.object(GarmCharm, "unit", new_callable=PropertyMock) as mock_unit,
     ):
+        mock_charm_state_cls.from_charm.return_value.desired_entities = entities
         mock_unit.return_value = MagicMock()
         order = MagicMock()
         order.attach_mock(mock_github_cls.return_value.reconcile, "github")
@@ -1035,100 +1036,6 @@ def test_build_desired_credentials_skips_non_numeric_ids():
     credentials = GarmCharm._build_desired_credentials(charm)
 
     assert credentials == []
-
-
-@pytest.mark.parametrize(
-    "unit_data, expected_type, expected_name",
-    [
-        (
-            {"org": "canonical", "github_app_id": "12345", "github_installation_id": "67890"},
-            "organization",
-            "canonical",
-        ),
-        (
-            {"repo": "canonical/runner", "github_app_id": "1", "github_installation_id": "2"},
-            "repository",
-            "canonical/runner",
-        ),
-    ],
-    ids=["org-entity", "repo-entity"],
-)
-def test_build_desired_entities_builds_entity_from_relation(
-    unit_data, expected_type, expected_name
-):
-    """
-    arrange: A charm whose configurator relation exposes a unit naming an org or repo plus the
-        GitHub App ids.
-    act: Call _build_desired_entities.
-    assert: One entity is built with the right type/name, bound to the app-<id>-<id> credential.
-    """
-    charm = _github_charm([unit_data])
-
-    entities = GarmCharm._build_desired_entities(charm)
-
-    assert len(entities) == 1
-    entity = entities[0]
-    assert entity.entity_type == expected_type
-    assert entity.entity_name == expected_name
-    assert entity.credentials_name == "app-{}-{}".format(
-        unit_data["github_app_id"], unit_data["github_installation_id"]
-    )
-
-
-def test_build_desired_entities_dedupes_by_name():
-    """
-    arrange: A charm whose configurator relation exposes two units naming the same org.
-    act: Call _build_desired_entities.
-    assert: The two units collapse to a single entity registration.
-    """
-    unit_data = {"org": "canonical", "github_app_id": "1", "github_installation_id": "2"}
-    charm = _github_charm([dict(unit_data), dict(unit_data)])
-
-    entities = GarmCharm._build_desired_entities(charm)
-
-    assert len(entities) == 1
-
-
-def test_build_desired_entities_distinguishes_org_and_repo_with_same_name():
-    """
-    arrange: A configurator relation exposing an org and a repo that share the same raw name.
-    act: Call _build_desired_entities.
-    assert: Two distinct entities are produced — dedup is by (type, name), not name alone, so the
-        org and repo do not collide.
-    """
-    charm = _github_charm(
-        [
-            {"org": "shared", "github_app_id": "1", "github_installation_id": "2"},
-            {"repo": "shared", "github_app_id": "1", "github_installation_id": "2"},
-        ]
-    )
-
-    entities = GarmCharm._build_desired_entities(charm)
-
-    assert {(e.entity_type, e.entity_name) for e in entities} == {
-        ("organization", "shared"),
-        ("repository", "shared"),
-    }
-
-
-@pytest.mark.parametrize(
-    "unit_data",
-    [
-        {"github_app_id": "1", "github_installation_id": "2"},
-        {"org": "canonical", "github_installation_id": "2"},
-        {"org": "canonical", "github_app_id": "x", "github_installation_id": "2"},
-    ],
-    ids=["no-org-or-repo", "missing-app-id", "non-numeric-id"],
-)
-def test_build_desired_entities_skips_incomplete_unit(unit_data):
-    """
-    arrange: A charm whose configurator unit lacks an entity name or valid GitHub App ids.
-    act: Call _build_desired_entities.
-    assert: No entity is built.
-    """
-    charm = _github_charm([unit_data])
-
-    assert GarmCharm._build_desired_entities(charm) == []
 
 
 def test_ensure_controller_urls_sets_urls_from_base_url():
